Fixed the code so the Receipt class will handle all the functions which uses the receipt. We add a function in the Controller class which makes a new instance from the receiptDTO.  in the receiptDTO class, the receiptDTO constractor has been changed.  the constructor of the Receipt class has been changed and new functions have been added (returnReceiptDTO, updatePayedCash).  In the class Sale, the function addItemsToListAndCalculatePrice parameters have been modified
Fixed the code so the Receipt class will handle all the functions which uses the receipt. We add a function in the Controller class which makes a new instance from the receiptDTO.  in the receiptDTO class, the receiptDTO constractor has been changed.  the constructor of the Receipt class has been changed and new functions have been added (returnReceiptDTO, updatePayedCash).  In the class Sale, the function addItemsToListAndCalculatePrice parameters have been modified.