throw the exception of ItemNotFount in Controller class instead of View class so the enterItem no longer returns null when not finding an item.   create the ExceptionForErrorInDatabase class. create the ExceptionNotFoundItem class. create the interface of the RegisterObserver. create the TotalRevenueViewRevenue view class deleted the if statment from the view and instead of it we just catch the error from the Controller. From the test we created the exception handling test which tests catching the NotfoundItem error exception
throw the exception of ItemNotFount in Controller class instead of View class so the enterItem no longer returns null when not finding an item.
create the ExceptionForErrorInDatabase class.
create the ExceptionNotFoundItem class.
create the interface of the RegisterObserver.
create the TotalRevenueViewRevenue view class
deleted the if statment from the view and instead of it we just catch the error from the Controller.
From the test we created the exception handling test which tests catching the NotfoundItem error exception.