in the View class we deleted the startUp call. we have added the call of the function addObserver in the View constructor with passing a new observer.  we change the name of the observer interface function to newSale.  the TotalRevenueViewRevenue implementetion of the RegistryObserver we have print out the total money of the registry which been added when notify the observer and prints the registryUpdates array size to show how many updates have been happen to the registry.  the new sale have been deleted from the function enterItem and moved to be a globel variable.  deleted the static from the inventorySystem
in the View class we deleted the startUp call. we have added the call of the function addObserver in the View constructor with passing a new observer.
we change the name of the observer interface function to newSale.
the TotalRevenueViewRevenue implementetion of the RegistryObserver we have print out the total money of the registry which been added when notify the observer and prints the registryUpdates array size to show how many updates have been happen to the registry.
the new sale have been deleted from the function enterItem and moved to be a globel variable.
deleted the static from the inventorySystem.