Deleted the parameter ArrayList<ReceiptItemsDTO> from the function "addItemsToListAndCalculatePrice" so we don't have to pass the array list from the controller.  added the two functions "returnReceiptItemsDTO" which we used in the class Controller. and the resetReceiptItemsDTO which we used when reseting the receipt. deletd the parmeters from the
Deleted the parameter ArrayList<ReceiptItemsDTO> from the function "addItemsToListAndCalculatePrice" so we don't have to pass the array list from the controller.
added the two functions "returnReceiptItemsDTO" which we used in the class Controller. and the resetReceiptItemsDTO which we used when reseting the receipt.
deletd the parmeters from the function createReceiptAndShowPrice.